+++ lofty/src/id3/v2/header.rs
@@ -1,7 +1,9 @@
-use crate::error::{Id3v2Error, Id3v2ErrorKind, Result};
+use crate::error::Result;
+use crate::id3::v2::error::{Id3v2Error, Id3v2ErrorKind};
 use crate::id3::v2::restrictions::TagRestrictions;
 use crate::id3::v2::util::synchsafe::SynchsafeInteger;
 use crate::macros::err;
+use std::fmt::Display;
 
 use std::io::Read;
 
@@ -16,6 +18,16 @@ pub enum Id3v2Version {
 	V3,
 	/// ID3v2.4
 	V4,
+}
+
+impl Display for Id3v2Version {
+	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+		match self {
+			Id3v2Version::V2 => write!(f, "ID3v2.2"),
+			Id3v2Version::V3 => write!(f, "ID3v2.3"),
+			Id3v2Version::V4 => write!(f, "ID3v2.4"),
+		}
+	}
 }
 
 /// Flags that apply to the entire tag