+++ lofty/src/id3/v2/items/mod.rs
@@ -1,6 +1,7 @@
 mod attached_picture_frame;
 mod audio_text_frame;
 mod binary_frame;
+mod chapter_frame;
 mod encapsulated_object;
 mod event_timing_codes_frame;
 mod extended_text_frame;
@@ -12,6 +13,7 @@ mod popularimeter;
 mod private_frame;
 mod relative_volume_adjustment_frame;
 mod sync_text;
+mod table_of_contents_frame;
 mod text_information_frame;
 mod timestamp_frame;
 mod unique_file_identifier;
@@ -20,6 +22,7 @@ mod url_link_frame;
 pub use attached_picture_frame::AttachedPictureFrame;
 pub use audio_text_frame::{AudioTextFrame, AudioTextFrameFlags, scramble};
 pub use binary_frame::BinaryFrame;
+pub use chapter_frame::ChapterFrame;
 pub use encapsulated_object::GeneralEncapsulatedObject;
 pub use event_timing_codes_frame::{Event, EventTimingCodesFrame, EventType};
 pub use extended_text_frame::ExtendedTextFrame;
@@ -33,6 +36,7 @@ pub use relative_volume_adjustment_frame::{
 	ChannelInformation, ChannelType, RelativeVolumeAdjustmentFrame,
 };
 pub use sync_text::{SyncTextContentType, SynchronizedTextFrame, TimestampFormat};
+pub use table_of_contents_frame::{ChapterTableOfContentsFrame, CtocFlags};
 pub use text_information_frame::TextInformationFrame;
 pub use timestamp_frame::TimestampFrame;
 pub use unique_file_identifier::UniqueFileIdentifierFrame;