+++ lofty/src/id3/v2/items/relative_volume_adjustment_frame.rs
@@ -1,5 +1,6 @@
 use crate::config::{ParsingMode, WriteOptions};
-use crate::error::{Id3v2Error, Id3v2ErrorKind, Result};
+use crate::error::Result;
+use crate::id3::v2::error::{Id3v2Error, Id3v2ErrorKind};
 use crate::id3::v2::{FrameFlags, FrameHeader, FrameId};
 use crate::macros::try_vec;
 use crate::util::text::{TextDecodeOptions, TextEncoding, decode_text};
@@ -243,10 +244,10 @@ impl<'a> RelativeVolumeAdjustmentFrame<'a> {
 	}
 }
 
-impl RelativeVolumeAdjustmentFrame<'static> {
-	pub(crate) fn downgrade(&self) -> RelativeVolumeAdjustmentFrame<'_> {
+impl RelativeVolumeAdjustmentFrame<'_> {
+	pub(crate) fn borrow(&self) -> RelativeVolumeAdjustmentFrame<'_> {
 		RelativeVolumeAdjustmentFrame {
-			header: self.header.downgrade(),
+			header: self.header.borrow(),
 			identification: Cow::Borrowed(&self.identification),
 			channels: Cow::Borrowed(&self.channels),
 		}