+++ lofty/src/id3/v2/mod.rs
@@ -7,6 +7,7 @@
 //! * [`Id3v2Tag`]
 //! * [`Frame`]
 
+pub mod error;
 mod frame;
 pub(crate) mod header;
 mod items;
@@ -26,6 +27,7 @@ pub use tag::Id3v2Tag;
 pub use items::*;
 
 pub use frame::header::{FrameHeader, FrameId};
+pub use frame::list::FrameList;
 pub use frame::{Frame, FrameFlags};
 
 pub use restrictions::{